Move TemporaryBuffers to internal and use it everywhere.
