+++ api/all/src/main/java/io/opentelemetry/api/internal/TemporaryBuffers.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.api.trace.propagation;
+package io.opentelemetry.api.internal;
 
 /**
  * {@link ThreadLocal} buffers for use when creating new derived objects such as {@link String}s.
@@ -12,7 +12,7 @@ package io.opentelemetry.api.trace.propagation;
  * should get a temporary buffer, fill it with data, and finish by converting into the derived
  * object within the same method to avoid multiple usages of the same buffer.
  */
-final class TemporaryBuffers {
+public final class TemporaryBuffers {
 
   private static final ThreadLocal<char[]> CHAR_ARRAY = new ThreadLocal<>();
 
@@ -22,7 +22,7 @@ final class TemporaryBuffers {
    * not be zeroed and may be larger than the requested size, you must make sure to fill the entire
    * content to the desired value and set the length explicitly when converting to a {@link String}.
    */
-  static char[] chars(int len) {
+  public static char[] chars(int len) {
     char[] buffer = CHAR_ARRAY.get();
     if (buffer == null || buffer.length < len) {
       buffer = new char[len];
